discord/pythonProject1/requirements.txt
crewai==0.30.11
crewai-tools==0.2.6
google-ai-generativelanguage==0.6.4
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==1.0.4
pandas
python-dotenv==1.0.1
IPython
streamlit